Terrarium + cherry
Fixed the cherry attach
Added Terrarium with terrain
