--- Assets/GrabberLearning.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: cfd1894c50594cf47ba9bcb56ec58fbb
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
