+++ Assets/ML-Agents/Timers/Scene Room_timers.json
@@ -1,1 +1,1 @@
-{"count":1,"self":728.3239936,"total":732.1619482,"children":{"AgentSendState":{"count":36327,"self":1.0616584,"total":1.3702147,"children":{"CollectObservations":{"count":3606,"self":0.023504499999999998,"total":0.023504499999999998,"children":null},"WriteActionMask":{"count":3606,"self":0.0060018,"total":0.0060018,"children":null},"RequestDecision":{"count":3606,"self":0.27904999999999996,"total":0.27904999999999996,"children":null}}},"DecideAction":{"count":36327,"self":2.2861957999999998,"total":2.3472052,"children":{"RayPerceptionSensor.Perceive":{"count":7212,"self":0.0610094,"total":0.0610094,"children":null}}},"AgentAct":{"count":36327,"self":0.1130209,"total":0.1130209,"children":null},"InitializeActuators":{"count":6,"self":0.0010004,"total":0.0010004,"children":null},"InitializeSensors":{"count":6,"self":0.0020009,"total":0.0060021,"children":{"CreateObservableSensors":{"count":6,"self":0.0040012,"total":0.0040012,"children":null}}}},"gauges":{"AntAgent.CumulativeReward":{"count":77,"max":33,"min":-10,"runningAverage":6.837663,"value":0,"weightedAverage":5.388454}},"metadata":{"timer_format_version":"0.1.0","start_time_seconds":"1716141403","unity_version":"2022.3.14f1","command_line_arguments":"T:\\Unity Hub\\Editor\\2022.3.14f1\\Editor\\Unity.exe -projectpath T:\\Github Projects\\SmartAnt -useHub -hubIPC -cloudEnvironment production -licensingIpc LicenseClient-Thomas -hubSessionId f2c42445-b814-4b3e-8550-d2cff2c547de -accessToken lHz_fgsuMq9Kek3n5aybYF1pv-09S-eoSYx52aa5cvQ012f","communication_protocol_version":"1.5.0","com.unity.ml-agents_version":"2.0.1","scene_name":"Scene Room","end_time_seconds":"1716142135"}}
+{"count":1,"self":101.57341439999999,"total":102.12237089999999,"children":{"AgentSendState":{"count":4827,"self":0.14364749999999998,"total":0.1586627,"children":{"CollectObservations":{"count":601,"self":0.0045051,"total":0.0045051,"children":null},"WriteActionMask":{"count":601,"self":0.0010007,"total":0.0010007,"children":null},"RequestDecision":{"count":601,"self":0.0095094,"total":0.0095094,"children":null}}},"DecideAction":{"count":4827,"self":0.3692773,"total":0.37428429999999996,"children":{"RayPerceptionSensor.Perceive":{"count":1202,"self":0.005007,"total":0.005007,"children":null}}},"AgentAct":{"count":4827,"self":0.011007099999999999,"total":0.011007099999999999,"children":null},"InitializeActuators":{"count":1,"self":0.0010008,"total":0.0010008,"children":null},"InitializeSensors":{"count":1,"self":0.0010012999999999999,"total":0.003503,"children":{"CreateObservableSensors":{"count":1,"self":0.0025017,"total":0.0025017,"children":null}}}},"gauges":{"AntAgent.CumulativeReward":{"count":6,"max":20.5,"min":-5,"runningAverage":5.5,"value":20.5,"weightedAverage":8.678711}},"metadata":{"timer_format_version":"0.1.0","start_time_seconds":"1716148637","unity_version":"2022.3.14f1","command_line_arguments":"C:\\Program Files\\Unity\\Hub\\Editor\\2022.3.14f1\\Editor\\Unity.exe -projectpath D:\\Vakken\\VRExperience\\SmartAnt -useHub -hubIPC -cloudEnvironment production -licensingIpc LicenseClient-yasso -hubSessionId f2e12414-7db5-4d15-a036-abac6c430835 -accessToken JbuKkMcbXlyZfA6qDYBj48T7S2CjGE_vWovEYhVeDlc008f","communication_protocol_version":"1.5.0","com.unity.ml-agents_version":"2.0.1","scene_name":"Scene Room","end_time_seconds":"1716148739"}}
