+++ Assets/Materials/PBR_Grass_Textures/Materials/grass_01.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b67cbd5fcc22cb6498017e64eb957671
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
