+++ Assets/Materials/TerrariumFloor.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: cb34d2d375cd8744cba037591670608b
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
