+++ Assets/Materials/TerrariumWalls.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3bf11dc3573eaa142aa5e9d39c1aeb7f
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
