--- Assets/ScoreTriggerLearning.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e21f5b7e5d632c94a83cf6d2d5f88093
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
