+++ Assets/Scripts/AnimalControllerLearning.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9c20db29b07caa443851c7637b470003
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
