+++ Assets/Terrains/Ground textures pack.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7374926c97525054fb0aa3a1ad1c2d9a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
