+++ Assets/Terrains/Ground textures pack/Ground 01.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 60d9638b2d0fc584eaa50ef2d7b99418
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
