+++ Assets/Terrains/Ground textures pack/Ground 01/Ground pattern 01.mat.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 748487d582b00d847bd57517f78d8e8f
+labels:
+- Ground
+- Nature
+- Pattern
+- Terrain
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
