+++ Assets/Terrains/New Terrain 1.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8db0f2b2372cad64996ca3774d8d5c26
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 15600000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
